Move AI to base object mngr.
/ GameObjectAI exist, hence this change.
