+++ game/world/managers/objects/ai/BasicCreatureAI.py
@@ -1,4 +1,4 @@
-from game.world.managers.objects.units.ai.CreatureAI import CreatureAI
+from game.world.managers.objects.ai.CreatureAI import CreatureAI
 from utils.constants.CustomCodes import Permits
